Rename soalDataTypes.jav to soalDataTypes.java
Java has 8 primitive data types; char, boolean, byte, short, int, long, float, and double. For this
exercise, we'll work with the primitives used to hold integer values (byte, short, int, and long):

This program determine which primitive data types are capable of properly storing that input
The try statement allows us to define a block of code to be tested for errors while it is being executed. 
The catch statement allows us to define a block of code to be executed, if an error occurs in the try block.